we need folderssssss

+++ peidashboard/dashboard/models.py
@@ -14,6 +14,8 @@ import os
 
 from datetime import datetime
 
+import time
+
 LEXERS = [item for item in get_all_lexers() if item[1]]
 LANGUAGE_CHOICES = sorted([(item[1][0], item[0]) for item in LEXERS])
 STYLE_CHOICES = sorted([(item, item) for item in get_all_styles()])
@@ -48,15 +50,24 @@ class Snippet(models.Model):
 class Document(models.Model):
     created = models.DateTimeField(auto_now_add=True)
     docfile = models.FileField(unique=True, upload_to='media')
+    docname = models.CharField(max_length=100, blank=True, default='')
     title = models.CharField(max_length=100, blank=True, default='')
     downloadlink = models.URLField(editable=False, default='')
+    # STATISTICS
+    #viewcount = models.PositiveIntegerField(default=0)
+    viewtime = models.PositiveIntegerField(default=0)
 
     class Meta:
         ordering = ['-created']
 
     def save(self, *args, **kwargs):
         self.downloadlink = 'http://'+settings.ALLOWED_HOSTS[0]+'/download/'+self.docfile.name.split('/')[-1]
+        self.docname = self.docfile.name.split('/')[-1]
         super(Document, self).save(*args, **kwargs)
+
+    @property
+    def pretty_viewtime(self):
+        return time.strftime('%H:%M:%S', time.gmtime(self.viewtime))
 
 class Space(models.Model):
     name = models.CharField(max_length=100, unique=True)
@@ -104,7 +115,13 @@ class AgentUpdate(models.Model):
     urltoken = models.CharField(max_length=100, unique=True)
     token_confirm = models.BooleanField(default=False)
     contentid = models.ForeignKey(Document, null=True, on_delete=models.SET_NULL)
+    contentname = models.CharField(max_length=100, blank=True, default='')
     content_confirm = models.BooleanField(default=False)
+
+    def save(self, *args, **kwargs):
+        doc = Document.objects.get(pk=self.contentid.id)
+        self.contentname = doc.docfile.name.split('/')[-1]
+        super(AgentUpdate, self).save(*args, **kwargs)
 
 class ContentProgram(models.Model):
     name = models.CharField(max_length=100, unique=True)
@@ -115,8 +132,10 @@ class ContentProgram(models.Model):
         ordering = ['start_date', 'name']
 
 class Crop(models.Model):
-    cropfile = models.FileField(unique=True, upload_to='crops')
+    cropfile = models.FileField(blank=True, upload_to='crops')
     content = models.TextField()
+    agentid = models.IntegerField()
+    interactiontime = models.IntegerField(blank=True, default=0)
 
     def save(self, *args, **kwargs):
         #print("Uploading crop...")
@@ -126,18 +145,26 @@ class Crop(models.Model):
 
         print("ola")
 
+        if Document.objects.filter(docname=self.content).count() != 0:
+            targetdoc = Document.objects.get(docname=self.content)
+            targetdoc.viewtime += self.interactiontime
+            targetdoc.save()
+
         # DELETE ACTUAL FILE
 
-        print(self.cropfile)
-        filename = os.path.dirname(os.path.abspath(__file__))+'/../media_cdn/'+str((self.cropfile))
-        print(filename)
-        os.remove(filename)
+        if self.cropfile != '':
+            print(self.cropfile)
+            filename = os.path.dirname(os.path.abspath(__file__))+'/../media_cdn/'+str((self.cropfile))
+            print(filename)
+            os.remove(filename)
+        else:
+            print("no file provided")
 
         # DELETE ENTRY FROM DB
 
         #print("Self", self.pk)
         crop = Crop.objects.get(id=int(self.pk))
         crop.delete()
-        #Crop.objects.all().delete()
+        Crop.objects.all().delete()
         #print("Deleting crop...")
         #print("Crops", Crop.objects.all())
